refactor: modularize BetTracker component
